Config and chat cleanup (#29)
* Cleanup

* Remove ChatUtils.color from arg of ValidationException.

+++ src/main/java/com/eternalcode/core/configuration/PluginConfiguration.java
@@ -34,7 +34,7 @@ public final class PluginConfiguration implements Serializable {
 
     @Description(" ")
     @Description("# Chat Settings")
-    public int chatSlowMode = 5;
+    public double chatSlowMode = 5.0;
     public boolean chatStatue = true;
 
     @Description(" ")
